pyproject.toml
[tool.poetry]
name = "ai-based-data-engineering"
version = "0.1.0"
description = ""
authors = ["Siddharth Adhikari <93575478+adhikarisidd17@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
openai = "^1.86.0"
gitpython = "^3.1.44"
pygithub = "^2.6.1"
jinja2 = "^3.1.6"
pydantic = "^2.11.5"
pytest = "^8.4.0"
dbt-core = "^1.9.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
